requirements.txt
# Python dependencies for PPTX extraction
python-pptx>=0.6.21
